Adds zoom to location button
Improves the user experience by providing a persistent zoom to location button, separate from settings.

This change adds a floating zoom button in the bottom-left corner that allows users to easily center the map on their current location. This provides quicker access to this functionality compared to accessing it through the settings panel. It also removes the dedicated zoom to location button from the settings panel.